Database Functionality Moved to Services
