+++ index.html
@@ -5,13 +5,15 @@
   <title>Hazard a Guess</title>
   <meta charset="utf-8" />
   <!-- JS code is executed after DOM is fully built -->
-  <script defer src = "/public/fetchMode.js">
+
 </head>
 
 <body>
   <h1 id="heading">Hazard a Guess Main Page</h1>
   <button type="button" id="helloButton">Say Hi :)</button>
   <button type="button" id="settingsButton">Settings</button>
-  </script>
+
+  <script defer src = "/public/fetchMode.js"> </script>
 </body>
+
 </html>
