+++ services/testDB.js
@@ -6,7 +6,7 @@ module.exports = {
     const pool = await get('default')
 
     // Get dark mode setting for specific user
-    const mode = await pool.request().query("SELECT Darkmode FROM User_Details WHERE Username = '"+ user+"'")
+    const mode = await pool.request().query("SELECT Darkmode FROM User_Details WHERE Username = '" + user + "'")
 
     if (mode.recordset[0].Darkmode == true) {
       return 'true'
@@ -27,5 +27,3 @@ module.exports = {
     ).catch(console.error)
   }
 }
-
-
